requirements.txt
astroid==2.12.13
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.7
dill==0.3.6
idna==3.4
isort==5.11.2
lazy-object-proxy==1.8.0
mccabe==0.7.0
numpy==1.23.4
pandas==1.5.0
platformdirs==2.6.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.8
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
twilio==7.15.1
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.2
